backend/pyproject.toml
[project]
name = "appwondo-backend"
version = "0.1.0"
description = "Appwondo Backend - Traditional Taekwon-Do Black Belt Center Torgau"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "PyJWT>=2.10.0",
    "bcrypt>=4.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    "slowapi>=0.1.9",
    "icalendar>=6.1.0",
    "python-dateutil>=2.9.0",
    "Pillow>=11.0.0",
    "python-multipart>=0.0.17",
    "httpx>=0.28.0",
    "websockets>=14.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
